Server - The start of the start
Started working on the server part. Having to re learn a lot of stuff. Does nothing right now other than post data to the server where the server then reads the data.

Next, setup a DB and have the server lookup a user id and show info on user.
Also I need to learn flask and make it so I can use a template or something otherwise this will get big and ugly...